Refactor code to restart for the second number
The calculator needs to keep the first input on the screen when the
operator is pressed, but once the second input is started the screen
needs to only show the second input numbers. The calculator was
appending the second input numbers on to the first so adding logic to
see if the first number and operator were present was added. This meant
moving the location of firstNumber and secondNumber assignment and
rewriting the operator event listener to reflect.