Add sign change functionality
The sign of the input can be changed from positive to negative and back
when inputting the number. If the equal sign has been pressed, the sign
of the result can be changed and used in subsequent operations. If a new
operation is started the sign of a zero cannot be changed.