Add input parsing for everything but decimals
The inputs need to be checked for overflow and infinity values. If the
inputs are greater than a certain length, depending on signage, the
input is replaced with an error message. The sign is an additional
character that consumes the 13th spot on the calculator and should not
reduce the length of the max possible input length.