Add logic to handing positive scientific notation
Javascript converts any number with more than 20 digits to scientific
notation. Scientific notation breaks the calculator and works despite
being too large of a number for a calculator that can only display 12
digits to handle. The code checks for positive scientific notation and
throws an overflow if any is detected.